# openapi.json
{"openapi":"3.1.0","info":{"title":"API FastAPI - Trabalho Final","description":"\n    API desenvolvida durante as aulas de API - FastAPI, \n    contendo endpoints de exemplo de uso com LLM, \n    API para extrair informações de relatos de conversas com clientes e \n    formatá-las para importação em um banco de dados.\n    ","termsOfService":"http://example.com/terms/","contact":{"name":"Carlos Miqui","url":"http://github.com/carlosmiqui/","email":"carlosmiqui@discente.ufg.br"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"},"version":"0.1"},"paths":{"/processar_relato":{"post":{"tags":["LLM"],"summary":"Processa um relato de conversa e retorna um objeto JSON formatado.","description":"Recebe um relato de conversa em texto, extrai informações relevantes e retorna um objeto JSON para ser importado em um banco de dados.","operationId":"processar_relato_endpoint_processar_relato_post","parameters":[{"name":"api_token","in":"query","required":true,"schema":{"type":"string","title":"Api Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelatoInput"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelatoConversa"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gerar_estrategia":{"post":{"tags":["LLM"],"summary":"Gera uma estratégia de conversão a partir dos dados do relato.","description":"Recebe um objeto JSON com os dados do relato e gera uma estratégia de conversão personalizada para o cliente.","operationId":"gerar_estrategia_endpoint_gerar_estrategia_post","parameters":[{"name":"api_token","in":"query","required":true,"schema":{"type":"string","title":"Api Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelatoConversa"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstrategiaConversao"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"EstrategiaConversao":{"properties":{"estrategia":{"type":"string","title":"Estrategia"}},"type":"object","required":["estrategia"],"title":"EstrategiaConversao"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RelatoConversa":{"properties":{"data_conversa":{"type":"string","title":"Data Conversa"},"identidade_cliente":{"type":"string","title":"Identidade Cliente"},"identidade_atendente":{"type":"string","title":"Identidade Atendente"},"resumo_assunto":{"type":"string","title":"Resumo Assunto"},"valor_orcamento":{"type":"number","title":"Valor Orcamento"},"forma_pagamento":{"type":"string","title":"Forma Pagamento"},"informacoes_organizadas":{"items":{"type":"string"},"type":"array","title":"Informacoes Organizadas"},"instrucoes_proximo_passo":{"type":"string","title":"Instrucoes Proximo Passo"}},"type":"object","required":["data_conversa","identidade_cliente","identidade_atendente","resumo_assunto","valor_orcamento","forma_pagamento","informacoes_organizadas","instrucoes_proximo_passo"],"title":"RelatoConversa"},"RelatoInput":{"properties":{"relato":{"type":"string","maxLength":10000,"title":"Relato","description":"Relato da conversa (texto formatado, sem limite de formatação)"}},"type":"object","required":["relato"],"title":"RelatoInput"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}